LayoutTests/fast/events/touch/ios/focus-textarea-on-touchstart.html

<!DOCTYPE html> <!-- webkit-test-runner [ useFlexibleViewport=true ] -->
<html>
<meta name="viewport" content="width=device-width, initial-scale=1">
<head>
<script src="../../../../resources/js-test.js" type="text/javascript"></script>
<script src="../../../../resources/ui-helper.js" type="text/javascript"></script>
<style>
body, html {
    margin: 0;
    width: 100%;
    height: 5000px;
}

textarea {
    width: 100%;
    height: 100px;
    position: absolute;
    top: calc(100% - 100px);
    left: 0;
    display: block;
    font-size: 20px;
    border: 1px solid tomato;
    border-radius: 4px;
    outline: none;
    box-sizing: border-box;
    padding: 6px;
}
</style>
<script>
jsTestIsAsync = true;

addEventListener("load", async () => {
    description("Verifies that the keyboard does not immediately dismiss after tapping on an editable element that is focused during `touchstart`. This test requires WebKitTestRunner.");

    const textarea = document.querySelector("textarea");
    textarea.addEventListener("touchstart", () => {
        textarea.focus();
    });

    let target = UIHelper.midPointOfRect(textarea.getBoundingClientRect());

    document.body.addEventListener("click", () => {
        // Empty click event handler to trigger a potential synthetic click over the body.
    });

    await UIHelper.setHardwareKeyboardAttached(false);

    willHideCountBeforeTapping = await UIHelper.keyboardWillHideCount();
    await UIHelper.sendEventStream(new UIHelper.EventStreamBuilder()
        .begin(target.x, target.y)
        .wait(0.2)
        .end()
        .takeResult());

    await UIHelper.waitForKeyboardToShow();

    // Also wait for the keyboard to begin dismissing, in the case where the bug reproduces.
    await UIHelper.delayFor(200);

    willHideCountAfterTapping = await UIHelper.keyboardWillHideCount();

    shouldBe("willHideCountBeforeTapping", "willHideCountAfterTapping");

    document.activeElement.blur();
    await UIHelper.waitForKeyboardToHide();
    testPassed("Dismissed keyboard");

    finishJSTest();
});
</script>
</head>
<body>
    <textarea>Tap here (slowly)</textarea>
</body>
</html>
